backend/services/synthetic_data.py
import google.generativeai as genai
import time

class SyntheticDataService:
    def __init__(self, api_key):
        genai.configure(api_key=api_key)
        # Using a model capable of image generation if available, or text-to-image API
        # For Hackathon 2026, assuming specific generation model or using standard Imagen via Gemini
        self.model = genai.GenerativeModel('models/gemini-2.0-flash-exp-image-generation') # Based on user's model list

    async def generate_defect_image(self, defect_type="swelling"):
        """
        Simulates the 'Poisson Image Editing' data augmentation from BatteryGPT paper
        by generating synthetic defective battery images.
        """
        prompt = f"""
        Generate a photorealistic close-up image of a lithium-ion pouch cell battery.
        The battery should exhibit clear signs of '{defect_type}'.
        The defect should be seamlessly blended into the surface (similar to Poisson editing).
        Technical, industrial lighting, high resolution.
        """
        
        try:
            # Note: The actual API call for image generation might differ slightly based on SDK version
            # This is a placeholder for the concept
            response = self.model.generate_content(prompt)
            # Assuming response contains image data or url
            return {"status": "generated", "description": f"Synthetic {defect_type} sample created.", "data": "simulated_image_data"}
        except Exception as e:
            print(f"Generation error: {e}")
            return {"error": str(e)}

# usage: synthetic_service = SyntheticDataService(os.getenv("GEMINI_API_KEY"))
