backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
google-generativeai
pydantic
pytest
httpx
numpy
chromadb
pypdf
matplotlib
pandas
